fix(blog): serve OG image from a route handler, not the opengraph-image convention
The `opengraph-image` file convention can't live inside the `[...slug]`
catch-all — Next appends a hidden metadata-id segment, leaving the catch-all
no longer last ("Catch-all must be the last part of the URL"), which broke the
build. Serve the image from a route handler under a distinct prefix instead:
`/blog/og/<slug>` (and `/zh/blog/og/<slug>`), force-static + generateStaticParams
so it bakes at build. Point openGraph.images / twitter:image at it via getSeo.
