+++ app/zh/blog/og/[...slug]/route.tsx
@@ -1,26 +1,21 @@
+// Dynamic blog OG image (zh) as a Route Handler. See app/blog/og/[...slug]
+// for why this is a route handler rather than the opengraph-image convention.
+
 import { generateBlogStaticParams } from 'utils/blog/generateBlogStaticParams';
 import { getBlogPost } from 'utils/blog/getBlogPost';
-import {
-  OG_ALT,
-  OG_CONTENT_TYPE,
-  OG_SIZE,
-  renderBlogOgImage,
-} from 'utils/og/blogOgImage';
+import { renderBlogOgImage } from 'utils/og/blogOgImage';
 
-export const alt = OG_ALT;
-export const size = OG_SIZE;
-export const contentType = OG_CONTENT_TYPE;
+export const dynamic = 'force-static';
 export const dynamicParams = false;
 
 export function generateStaticParams() {
   return generateBlogStaticParams('zh');
 }
 
-export default async function Image({
-  params,
-}: {
-  params: { slug: string[] };
-}) {
+export async function GET(
+  _request: Request,
+  { params }: { params: { slug: string[] } },
+) {
   const slugPath = params.slug.join('/');
   const { post } = await getBlogPost('zh', slugPath);
   return renderBlogOgImage({
